+++ src/parser/enrichers/class/sorcerer/InnateSorcery.mjs
@@ -3,13 +3,16 @@ import DDBEnricherData from "../../data/DDBEnricherData.mjs";
 
 export default class InnateSorcery extends DDBEnricherData {
 
-  get useDefaultAdditionalActivities() {
-    return true;
+  get activity() {
+    return {
+      name: "Innate Sorcery",
+    };
   }
 
   get effects() {
     return [
       {
+        activityMatch: "Innate Sorcery",
         changes: [
           DDBEnricherData.ChangeHelper.unsignedAddChange("1", 20, "system.bonuses.spell.dc"),
         ],